refactor: split the mapping result builder by mapping direction
MappingResultBuilder served input and output mappings from one class,
annotating its own fields "input mappings only" and "output mappings
only" -- a class that has to say which half of itself applies is two
classes. Each instance also carried a no-op resolver for the mode it was
not in.

Split it into a sealed interface with one implementation per direction.
Neither declares a hook for the other's behaviour, because the split
removes work rather than relocating it: the input builder can never
poison a level, so its put and getVariable lose those branches entirely,
and the output builder never layers on read. What is genuinely shared is
the msgpack representation -- the level record, the nil sentinel, and the
serializer -- which moves to a package-private helper with no notion of
direction.

Pure restructure: the same 515 tests pass, and the unit tests move
between two classes without an assertion changing.
