+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/deployment/model/transformer/VariableMappingTransformer.java
@@ -39,11 +39,13 @@ import java.util.stream.Collectors;
  * {@link OutputMapping} and {@code BpmnVariableMappingBehavior}).
  *
  * <p>Input and output mappings differ in how a nested target relates to the existing scope variable
- * at runtime, and both differences are handled by {@code MappingResultBuilder}. An output mapping's
- * result is merged into that variable if it is a JSON object, at every nesting level, so the merged
- * value is what gets written. An input mapping's result is not merged — it is written as-is — but a
- * later source reading the target's root name sees the mapped keys layered over the value the scope
- * chain resolves, so a mapping shadows only the keys it defines.
+ * at runtime, and each difference is handled by its own {@code MappingResultBuilder}
+ * implementation. An output mapping's result is merged into that variable if it is a JSON object,
+ * at every nesting level, so the merged value is what gets written ({@code
+ * OutputMappingResultBuilder}). An input mapping's result is not merged — it is written as-is — but
+ * a later source reading the target's root name sees the mapped keys layered over the value the
+ * scope chain resolves, so a mapping shadows only the keys it defines ({@code
+ * InputMappingResultBuilder}).
  */
 public final class VariableMappingTransformer {
 
@@ -92,7 +94,7 @@ public final class VariableMappingTransformer {
    * Transforms the output mappings, keeping each mapping as its own source expression plus target
    * path so they can be evaluated one by one in modeling order at runtime. A nested target merges
    * with the existing scope value at every path level at runtime (see {@code
-   * MappingResultBuilder}).
+   * OutputMappingResultBuilder}).
    */
   public List<OutputMapping> transformOutputMappings(
       final Collection<? extends ZeebeMapping> outputMappings,
