+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/variable/InputMappingResultBuilderTest.java
@@ -13,14 +13,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.junit.jupiter.api.Test;
 
-final class MappingResultBuilderTest {
+final class InputMappingResultBuilderTest {
 
-  private final MappingResultBuilder builder = MappingResultBuilder.forInputMappings(name -> null);
+  private final InputMappingResultBuilder builder = new InputMappingResultBuilder(name -> null);
 
   @Test
   void shouldBuildEmptyDocument() {
@@ -138,113 +137,10 @@ final class MappingResultBuilderTest {
   }
 
   @Test
-  void shouldSeedNestedTargetFromScopeValue() {
-    // given: scope has a = {'c': 2}
-    final var builder =
-        MappingResultBuilder.forOutputMappings(
-            path -> path.equals(List.of("a")) ? asMsgPack("{'c': 2}") : null);
-
-    // when
-    builder.put(List.of("a", "b"), asMsgPack("1"));
-
-    // then: existing sibling 'c' survives the merge
-    assertEquality(builder.toDocument(), "{'a': {'b': 1, 'c': 2}}");
-  }
-
-  @Test
-  void shouldSeedEveryPathLevelFromScopeValue() {
-    // given: scope has a = {'b': {'d': 2}, 'e': 3}
-    final var scope =
-        Map.of(
-            List.of("a"), asMsgPack("{'b': {'d': 2}, 'e': 3}"),
-            List.of("a", "b"), asMsgPack("{'d': 2}"));
-    final var builder = MappingResultBuilder.forOutputMappings(scope::get);
-
-    // when
-    builder.put(List.of("a", "b", "c"), asMsgPack("1"));
-
-    // then: siblings survive at both levels
-    assertEquality(builder.toDocument(), "{'a': {'b': {'c': 1, 'd': 2}, 'e': 3}}");
-  }
-
-  @Test
-  void shouldPoisonLevelWhenScopeValueIsNotAMap() {
-    // given: scope has a = 5
-    final var builder =
-        MappingResultBuilder.forOutputMappings(
-            path -> path.equals(List.of("a")) ? asMsgPack("5") : null);
-
-    // when
-    builder.put(List.of("a", "b"), asMsgPack("1"));
-
-    // then: the level is null, like FEEL's context merge(5, {...})
-    assertEquality(builder.toDocument(), "{'a': null}");
-    assertEquality(builder.getVariable("a"), "null");
-  }
-
-  @Test
-  void shouldDiscardFurtherPutsUnderPoisonedLevel() {
-    final var builder =
-        MappingResultBuilder.forOutputMappings(
-            path -> path.equals(List.of("a")) ? asMsgPack("5") : null);
-
-    // when
-    builder.put(List.of("a", "b"), asMsgPack("1"));
-    builder.put(List.of("a", "c"), asMsgPack("2"));
-
-    // then
-    assertEquality(builder.toDocument(), "{'a': null}");
-  }
-
-  @Test
-  void shouldReplacePoisonedLevelWithPlainTargetPut() {
-    final var builder =
-        MappingResultBuilder.forOutputMappings(
-            path -> path.equals(List.of("a")) ? asMsgPack("5") : null);
-
-    // when: a nested put poisons 'a', then a plain put replaces it
-    builder.put(List.of("a", "b"), asMsgPack("1"));
-    builder.put(List.of("a"), asMsgPack("7"));
-
-    // then
-    assertEquality(builder.toDocument(), "{'a': 7}");
-  }
-
-  @Test
-  void shouldReseedFromScopeWhenNestedPutFollowsPlainPut() {
-    // given: scope a = {'c': 2}; a plain put stored a value buffer for 'a'
-    final var builder =
-        MappingResultBuilder.forOutputMappings(
-            path -> path.equals(List.of("a")) ? asMsgPack("{'c': 2}") : null);
-    builder.put(List.of("a"), asMsgPack("{'z': 9}"));
-
-    // when: the shape flips back to a map — the plain mapping's value is discarded and the fresh
-    // level is seeded from the SCOPE value, not the previously mapped value
-    builder.put(List.of("a", "b"), asMsgPack("1"));
-
-    // then
-    assertEquality(builder.toDocument(), "{'a': {'b': 1, 'c': 2}}");
-  }
-
-  @Test
-  void shouldSeedFreshMapWhenScopeValueIsNil() {
-    // given: scope has a = null — FEEL's `if (a != null)` takes the else branch
-    final var builder =
-        MappingResultBuilder.forOutputMappings(
-            path -> path.equals(List.of("a")) ? asMsgPack("null") : null);
-
-    // when
-    builder.put(List.of("a", "b"), asMsgPack("1"));
-
-    // then
-    assertEquality(builder.toDocument(), "{'a': {'b': 1}}");
-  }
-
-  @Test
   void shouldLayerNestedLevelOverShadowedValue() {
     // given: the scope chain resolves x to {'a': 1, 'b': 2}
     final var builder =
-        MappingResultBuilder.forInputMappings(
+        new InputMappingResultBuilder(
             name -> name.equals("x") ? asMsgPack("{'a': 1, 'b': 2}") : null);
 
     // when: only x.a is mapped
@@ -260,7 +156,7 @@ final class MappingResultBuilderTest {
   void shouldLayerEveryLevelOfANestedLevelOverShadowedValue() {
     // given: a shadowed value nested as deeply as the target path
     final var builder =
-        MappingResultBuilder.forInputMappings(
+        new InputMappingResultBuilder(
             name -> name.equals("x") ? asMsgPack("{'a': {'b': 1, 'c': 2}, 'd': 4}") : null);
 
     // when
@@ -275,7 +171,7 @@ final class MappingResultBuilderTest {
   void shouldNotLayerAValueAssignedToTheWholeName() {
     // given
     final var builder =
-        MappingResultBuilder.forInputMappings(
+        new InputMappingResultBuilder(
             name -> name.equals("x") ? asMsgPack("{'a': 1, 'b': 2}") : null);
 
     // when: x is assigned as a whole, not at a path inside it
@@ -289,7 +185,7 @@ final class MappingResultBuilderTest {
   void shouldNotLayerALevelRecreatedAfterTheWholeNameWasAssigned() {
     // given: x was assigned whole, so the value it shadowed is already gone
     final var builder =
-        MappingResultBuilder.forInputMappings(
+        new InputMappingResultBuilder(
             name -> name.equals("x") ? asMsgPack("{'a': 1, 'b': 2}") : null);
     builder.put(List.of("x"), asMsgPack("1"));
 
@@ -304,7 +200,7 @@ final class MappingResultBuilderTest {
   void shouldNotLayerANestedLevelRecreatedAfterThatLevelWasAssigned() {
     // given: the case above one level down — this is why the flag is per level, not per name
     final var builder =
-        MappingResultBuilder.forInputMappings(
+        new InputMappingResultBuilder(
             name -> name.equals("x") ? asMsgPack("{'a': {'b': 1}, 'd': 4}") : null);
     builder.put(List.of("x", "a"), asMsgPack("1"));
 
@@ -319,7 +215,7 @@ final class MappingResultBuilderTest {
   void shouldNotLayerWhenTheShadowedValueIsNotAMap() {
     // given
     final var builder =
-        MappingResultBuilder.forInputMappings(name -> name.equals("x") ? asMsgPack("5") : null);
+        new InputMappingResultBuilder(name -> name.equals("x") ? asMsgPack("5") : null);
 
     // when
     builder.put(List.of("x", "a"), asMsgPack("1"));
@@ -332,7 +228,7 @@ final class MappingResultBuilderTest {
   @Test
   void shouldNotLayerWhenNothingShadowsTheName() {
     // given
-    final var builder = MappingResultBuilder.forInputMappings(name -> null);
+    final var builder = new InputMappingResultBuilder(name -> null);
 
     // when
     builder.put(List.of("x", "a"), asMsgPack("1"));
@@ -345,7 +241,7 @@ final class MappingResultBuilderTest {
   void shouldNotLetALayeredReadLeakIntoTheDocument() {
     // given
     final var builder =
-        MappingResultBuilder.forInputMappings(
+        new InputMappingResultBuilder(
             name -> name.equals("x") ? asMsgPack("{'a': 1, 'b': 2}") : null);
     builder.put(List.of("x", "a"), asMsgPack("3"));
 
@@ -362,8 +258,7 @@ final class MappingResultBuilderTest {
     // given — a 50 000-segment target path whose shadowed value is a map at the top level, so the
     // layering traversal descends the whole structure
     final var builder =
-        MappingResultBuilder.forInputMappings(
-            name -> name.equals("a") ? asMsgPack("{'a': 1}") : null);
+        new InputMappingResultBuilder(name -> name.equals("a") ? asMsgPack("{'a': 1}") : null);
     final var targetPath =
         IntStream.range(0, 50_000).mapToObj(i -> "a").collect(Collectors.toList());
     builder.put(targetPath, asMsgPack("1"));
